Revert "Revert "Bug fixes and new small features on user page""
