+++ .gitignore
@@ -7,3 +7,4 @@ bookly_project/backend/uploads/profiles/*
 bookly_project/backend/uploads/salons/*
 !bookly_project/backend/uploads/profiles/.gitkeep
 !bookly_project/backend/uploads/salons/.gitkeep
+bookly_project/CLAUDE.md
