+++ bookly_project/frontend/src/modules/Provider/ProvDash.jsx
@@ -1,10 +1,9 @@
-import { useState, useRef, useEffect, useCallback } from 'react';
-import { createPortal } from 'react-dom';
+import { useState, useRef, useEffect } from 'react';
 import Logo from '../../modules/Logo';
 import { useNotification } from '../../components/NotificationContext';
 import { useNavigate } from 'react-router-dom';
 import { useAuth } from '../auth/auth';
-import { authApi, logout } from '../auth/auth';
+import { authApi, logout, getUserFromToken } from '../auth/auth';
 import OverviewIcon from '../../icons/OverviewIcon';
 import CalendarIcon from '../../icons/CalendarIcon';
 import ServicesIcon from '../../icons/ServicesIcon';
@@ -12,1471 +11,13 @@ import SalonIcon from '../../icons/SalonIcon';
 import HourIcon from '../../icons/HourIcon';
 import SalonManagement from './SalonManagement';
 import AvailabilityManagement from './AvailabilityManagement';
-import { getUserFromToken } from '../auth/auth';
-import TimeSpinner from '../../components/TimeSpinner';
-import TimeBlockModal from './TimeBlockModal';
-import { timeBlocksService } from '../../services/timeBlocksService';
-import CurrencyIcon from '../../icons/CurrencyIcon';
-import UsersIcon from '../../icons/UsersIcon';
-import LeftArrowIcon from '../../icons/LeftArrowIcon';
-import RightArrowIcon from '../../icons/RightArrowIcon';
-import CloseIcon from '../../icons/CloseIcon';
-import UserIcon from '../../icons/UserIcon';
-import ChatBubbleIcon from '../../icons/ChatBubbleIcon';
-import TrashIcon from '../../icons/TrashIcon';
-import PencilIcon from '../../icons/PencilIcon';
-import WarningIcon from '../../icons/WarningIcon';
-import SettingsIcon from '../../icons/SettingsIcon';
-import LogoutIcon from '../../icons/LogoutIcon';
-import CheckCircleIcon from '../../icons/CheckCircleIcon';
-import LockIcon from '../../icons/LockIcon';
-
-// Section Components
-const OverviewSection = () => {
-    const [statistics, setStatistics] = useState({
-        todayAppointments: 0,
-        weeklyRevenue: 0,
-        newCustomers: 0,
-        upcomingAppointments: []
-    });
-    const [loading, setLoading] = useState(true);
-
-    useEffect(() => {
-        fetchStatistics();
-    }, []);
-
-    const fetchStatistics = async () => {
-        try {
-            setLoading(true);
-            const response = await authApi.get('/api/provider/calendar/statistics');
-            const data = await response.json();
-            if (data.success) {
-                setStatistics(data.statistics);
-            }
-        } catch (error) {
-            console.error('Error fetching statistics:', error);
-        } finally {
-            setLoading(false);
-        }
-    };
-
-    const formatPrice = (price) => {
-        return new Intl.NumberFormat('hu-HU').format(price);
-    };
-
-    const formatDate = (dateString) => {
-        return new Date(dateString).toLocaleDateString('hu-HU', {
-            month: 'short',
-            day: 'numeric',
-            hour: '2-digit',
-            minute: '2-digit',
-
-        });
-    };
-
-    const getStatusColor = (status) => {
-        switch (status) {
-            case 'scheduled': return 'bg-green-100 text-green-700';
-            case 'completed': return 'bg-blue-100 text-blue-700';
-            case 'canceled': return 'bg-red-100 text-red-700';
-            default: return 'bg-gray-100 text-gray-700';
-        }
-    };
-
-    const getStatusText = (status) => {
-        switch (status) {
-            case 'scheduled': return 'Megerősítve';
-            case 'completed': return 'Teljesítve';
-            case 'canceled': return 'Törölve';
-            default: return status;
-        }
-    };
-
-    const getInitials = (name) => {
-        if (!name) return '?';
-        const parts = name.split(' ');
-        if (parts.length >= 2) {
-            return parts[0][0] + parts[1][0];
-        }
-        return name.substring(0, 2).toUpperCase();
-    };
-
-    if (loading) {
-        return (
-            <div className="flex items-center justify-center py-12">
-                <div className="animate-spin rounded-full h-12 w-12 border-b-2 border-dark-blue"></div>
-            </div>
-        );
-    }
-
-    return (
-        <div className="space-y-6">
-            <h2 className="text-2xl font-bold text-dark-blue">Áttekintés</h2>
-            <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-6">
-                <div className="bg-white/40 backdrop-blur-md p-6 rounded-2xl shadow-lg border border-white/50 hover:shadow-xl transition-shadow">
-                    <div className="flex items-center justify-between">
-                        <div>
-                            <h3 className="text-gray-600 font-medium">Mai Foglalások</h3>
-                            <p className="text-4xl font-bold text-dark-blue mt-2">{statistics.todayAppointments}</p>
-                        </div>
-                        <div className="w-12 h-12 bg-blue-100 rounded-full flex items-center justify-center">
-                            <CalendarIcon className="text-dark-blue" />
-                        </div>
-                    </div>
-                </div>
-                <div className="bg-white/40 backdrop-blur-md p-6 rounded-2xl shadow-lg border border-white/50 hover:shadow-xl transition-shadow">
-                    <div className="flex items-center justify-between">
-                        <div>
-                            <h3 className="text-gray-600 font-medium">Heti Bevétel</h3>
-                            <p className="text-4xl font-bold text-dark-blue mt-2">{formatPrice(statistics.weeklyRevenue)} Ft</p>
-                        </div>
-                        <div className="w-12 h-12 bg-green-100 rounded-full flex items-center justify-center">
-                            <CurrencyIcon className="w-6 h-6 text-green-600" />
-                        </div>
-                    </div>
-                </div>
-                <div className="bg-white/40 backdrop-blur-md p-6 rounded-2xl shadow-lg border border-white/50 hover:shadow-xl transition-shadow">
-                    <div className="flex items-center justify-between">
-                        <div>
-                            <h3 className="text-gray-600 font-medium">Új Ügyfelek</h3>
-                            <p className="text-4xl font-bold text-dark-blue mt-2">{statistics.newCustomers}</p>
-                        </div>
-                        <div className="w-12 h-12 bg-purple-100 rounded-full flex items-center justify-center">
-                            <UsersIcon className="w-6 h-6 text-purple-600" />
-                        </div>
-                    </div>
-                </div>
-            </div>
-            
-            <div className="bg-white/40 backdrop-blur-md p-6 rounded-2xl shadow-lg border border-white/50">
-                <h3 className="text-xl font-bold text-dark-blue mb-4">Következő Időpontok</h3>
-                {statistics.upcomingAppointments.length === 0 ? (
-                    <div className="text-center py-8 text-gray-500">
-                        <CalendarIcon className="w-12 h-12 mx-auto mb-2 opacity-50" />
-                        <p>Nincsenek közelgő időpontok</p>
-                    </div>
-                ) : (
-                    <div className="space-y-3">
-                        {statistics.upcomingAppointments.map((appointment) => (
-                            <div key={appointment.id} className="flex items-center justify-between p-3 bg-white/50 rounded-lg hover:bg-white/70 transition-colors">
-                                <div className="flex items-center gap-4">
-                                    <div className="w-10 h-10 rounded-full bg-gradient-to-br from-dark-blue to-blue-500 flex items-center justify-center text-white font-bold text-sm shadow-md">
-                                        {getInitials(appointment.user_name)}
-                                    </div>
-                                    <div>
-                                        <p className="font-semibold text-gray-800">{appointment.user_name}</p>
-                                        <p className="text-sm text-gray-600">
-                                            {appointment.service_name} • {formatDate(appointment.appointment_start)}
-                                        </p>
-                                    </div>
-                                </div>
-                                <div className="flex items-center gap-3">
-                                    <span className="font-semibold text-dark-blue">
-                                        {formatPrice(appointment.price)} Ft
-                                    </span>
-                                    <span className={`px-3 py-1 rounded-full text-xs font-medium ${getStatusColor(appointment.status)}`}>
-                                        {getStatusText(appointment.status)}
-                                    </span>
-                                </div>
-                            </div>
-                        ))}
-                    </div>
-                )}
-            </div>
-        </div>
-    );
-};
-
-const CalendarSection = () => {
-    const [currentDate, setCurrentDate] = useState(new Date());
-    const [selectedDate, setSelectedDate] = useState(new Date());
-    const [appointments, setAppointments] = useState([]);
-    const [loading, setLoading] = useState(true);
-    const [selectedAppointment, setSelectedAppointment] = useState(null);
-    const [showModal, setShowModal] = useState(false);
-    const [deleteLoading, setDeleteLoading] = useState(false);
-    
-    // Create appointment modal
-    const [showCreateModal, setShowCreateModal] = useState(false);
-    const [services, setServices] = useState([]);
-    const [saving, setSaving] = useState(false);
-    const [createFormData, setCreateFormData] = useState({
-        is_guest: false,
-        user_email: '',
-        user_name: '',
-        user_phone: '',
-        service_id: '',
-        appointment_date: '',
-        appointment_time: '',
-        comment: ''
-    });
-    
-    // Working hours state (fetched from database)
-    const [workingHours, setWorkingHours] = useState({ openingHour: 8, closingHour: 20 });
-
-    // Time blocks state
-    const [timeBlocks, setTimeBlocks] = useState([]);
-    const [showTimeBlockModal, setShowTimeBlockModal] = useState(false);
-    const [selectedTimeBlock, setSelectedTimeBlock] = useState(null);
-    const { showToast } = useNotification();
-
-    // Derived timeline constants
-    const START_HOUR = workingHours.openingHour;
-    const END_HOUR = workingHours.closingHour;
-    const MINUTES_PER_PIXEL = 1; // 1 minute = 1 pixel for better precision
-    const TOTAL_MINUTES = (END_HOUR - START_HOUR) * 60;
-    
-    // Hours for the day view (only show hours where appointments can start)
-    const hours = Array.from({ length: END_HOUR - START_HOUR }, (_, i) => i + START_HOUR);
-
-    // Get start and end of current month for fetching appointments
-    const getMonthRange = (date) => {
-        const start = new Date(date.getFullYear(), date.getMonth(), 1);
-        const end = new Date(date.getFullYear(), date.getMonth() + 1, 0);
-        return {
-            startDate: start.toISOString().split('T')[0],
-            endDate: end.toISOString().split('T')[0]
-        };
-    };
-
-    // Fetch working hours
-    const fetchWorkingHours = useCallback(async () => {
-        try {
-            const response = await authApi.get('/api/provider/calendar/working-hours');
-            const data = await response.json();
-            if (data.success) {
-                setWorkingHours({
-                    openingHour: data.openingHour,
-                    closingHour: data.closingHour
-                });
-            }
-        } catch (error) {
-            console.error('Error fetching working hours:', error);
-        }
-    }, []);
-
-    // Fetch appointments
-    const fetchAppointments = useCallback(async () => {
-        setLoading(true);
-        try {
-            const { startDate, endDate } = getMonthRange(currentDate);
-            const response = await authApi.get(`/api/provider/calendar/appointments?startDate=${startDate}&endDate=${endDate}`);
-            const data = await response.json();
-            if (data.success) {
-                setAppointments(data.appointments);
-            }
-        } catch (error) {
-            console.error('Error fetching appointments:', error);
-        } finally {
-            setLoading(false);
-        }
-    }, [currentDate]);
-
-    // Fetch time blocks for the current month
-    const fetchTimeBlocks = useCallback(async () => {
-        try {
-            const { startDate, endDate } = getMonthRange(currentDate);
-            const data = await timeBlocksService.getTimeBlocks(startDate, endDate);
-            if (data.success) {
-                setTimeBlocks(data.timeBlocks);
-            }
-        } catch (error) {
-            console.error('Error fetching time blocks:', error);
-        }
-    }, [currentDate]);
-
-    // Fetch working hours on mount
-    useEffect(() => {
-        fetchWorkingHours();
-        fetchServices();
-    }, [fetchWorkingHours]);
-
-    useEffect(() => {
-        fetchAppointments();
-        fetchTimeBlocks();
-    }, [fetchAppointments, fetchTimeBlocks]);
-
-    // Fetch services for dropdown
-    const fetchServices = async () => {
-        try {
-            const response = await authApi.get('/api/provider/calendar/services');
-            const data = await response.json();
-            if (data.success) {
-                setServices(data.services.filter(s => s.status === 'available'));
-            }
-        } catch (error) {
-            console.error('Error fetching services:', error);
-        }
-    };
-
-    // Handle opening create modal
-    const handleOpenCreateModal = () => {
-        // Create a Date object for the default time (opening hour)
-        const defaultTime = new Date();
-        defaultTime.setHours(workingHours.openingHour, 0, 0, 0);
-        
-        setCreateFormData({
-            is_guest: false,
-            user_email: '',
-            user_name: '',
-            user_phone: '',
-            service_id: '',
-            appointment_date: selectedDate.toISOString().split('T')[0],
-            appointment_time: `${workingHours.openingHour.toString().padStart(2, '0')}:00`,
-            comment: ''
-        });
-        setShowCreateModal(true);
-    };
-
-    // Handle create appointment
-    const handleCreateAppointment = async () => {
-        // Validation
-        if (!createFormData.user_name || !createFormData.service_id) {
-            showToast('Név és szolgáltatás megadása kötelező!', 'warning');
-            return;
-        }
-
-        if (!createFormData.is_guest && !createFormData.user_email) {
-            showToast('Regisztrált felhasználóhoz email cím szükséges!', 'warning');
-            return;
-        }
-
-        if (createFormData.is_guest && !createFormData.user_email && !createFormData.user_phone) {
-            showToast('Vendég foglaláshoz legalább email vagy telefonszám szükséges!', 'warning');
-            return;
-        }
-
-        // Validate appointment time is within salon hours
-        const [hours, minutes] = createFormData.appointment_time.split(':').map(Number);
-        if (hours < workingHours.openingHour || hours >= workingHours.closingHour) {
-            showToast('Az időpont a szalon nyitvatartási idején kívül esik!', 'warning');
-            return;
-        }
-
-        try {
-            setSaving(true);
-            const response = await authApi.post('/api/provider/calendar/appointments', createFormData);
-            const data = await response.json();
-            
-            if (data.success) {
-                setShowCreateModal(false);
-                fetchAppointments();
-            } else {
-                showToast(data.message || 'Hiba történt a foglalás létrehozásakor', 'error');
-            }
-        } catch (error) {
-            console.error('Create appointment error:', error);
-            showToast('Hiba történt a foglalás létrehozásakor', 'error');
-        } finally {
-            setSaving(false);
-        }
-    };
-
-    // Get appointments for selected date
-    const getAppointmentsForDate = (date) => {
-        return appointments.filter(apt => {
-            const aptDate = new Date(apt.appointment_start);
-            return aptDate.toDateString() === date.toDateString();
-        });
-    };
-
-    // Get time blocks for selected date
-    const getTimeBlocksForDate = (date) => {
-        return timeBlocks.filter(block => {
-            const blockDate = new Date(block.start_datetime);
-            return blockDate.toDateString() === date.toDateString();
-        });
-    };
-
-    // Calculate time block position and height for timeline (same logic as appointments)
-    const getTimeBlockStyle = (block) => {
-        const start = new Date(block.start_datetime);
-        const end = new Date(block.end_datetime);
-        
-        const startMinutes = start.getHours() * 60 + start.getMinutes();
-        const endMinutes = end.getHours() * 60 + end.getMinutes();
-        
-        const topOffset = startMinutes - (START_HOUR * 60);
-        const duration = endMinutes - startMinutes;
-        
-        return {
-            top: `${topOffset * MINUTES_PER_PIXEL}px`,
-            height: `${Math.max(duration * MINUTES_PER_PIXEL, 20)}px`,
-        };
-    };
-
-    // Handle time block click
-    const handleTimeBlockClick = (block) => {
-        setSelectedTimeBlock(block);
-        setShowTimeBlockModal(true);
-    };
-
-    // Handle time block saved
-    const handleTimeBlockSaved = () => {
-        fetchTimeBlocks();
-        showToast('Szünet sikeresen mentve');
-    };
-
-    // Calculate appointment position and height for timeline
-    const getAppointmentStyle = (apt) => {
-        const start = new Date(apt.appointment_start);
-        const end = new Date(apt.appointment_end);
-        
-        const startMinutes = start.getHours() * 60 + start.getMinutes();
-        const endMinutes = end.getHours() * 60 + end.getMinutes();
-        
-        const topOffset = startMinutes - (START_HOUR * 60);
-        const duration = endMinutes - startMinutes;
-        
-        return {
-            top: `${topOffset * MINUTES_PER_PIXEL}px`,
-            height: `${Math.max(duration * MINUTES_PER_PIXEL, 30)}px`, // Minimum 30px height
-        };
-    };
-
-    // Calendar navigation
-    const prevMonth = () => {
-        setCurrentDate(new Date(currentDate.getFullYear(), currentDate.getMonth() - 1, 1));
-    };
-
-    const nextMonth = () => {
-        setCurrentDate(new Date(currentDate.getFullYear(), currentDate.getMonth() + 1, 1));
-    };
-
-    // Generate calendar days
-    const generateCalendarDays = () => {
-        const year = currentDate.getFullYear();
-        const month = currentDate.getMonth();
-        const firstDay = new Date(year, month, 1);
-        const lastDay = new Date(year, month + 1, 0);
-        const startDayOfWeek = firstDay.getDay() === 0 ? 6 : firstDay.getDay() - 1; // Monday = 0
-        const daysInMonth = lastDay.getDate();
-        
-        const days = [];
-        
-        // Empty cells for days before month starts
-        for (let i = 0; i < startDayOfWeek; i++) {
-            days.push(null);
-        }
-        
-        // Days of the month
-        for (let day = 1; day <= daysInMonth; day++) {
-            days.push(new Date(year, month, day));
-        }
-        
-        return days;
-    };
-
-    // Total slots based on working hours
-    const TOTAL_SLOTS = END_HOUR - START_HOUR;
-
-    // Get booking status for a date: 'empty' | 'available' | 'full'
-    const getBookingStatus = (date) => {
-        if (!date) return 'empty';
-        
-        const dayAppointments = appointments.filter(apt => {
-            const aptDate = new Date(apt.appointment_start);
-            return aptDate.toDateString() === date.toDateString() && apt.status !== 'canceled';
-        });
-
-        if (dayAppointments.length === 0) return 'empty';
-        if (dayAppointments.length >= TOTAL_SLOTS) return 'full';
-        return 'available';
-    };
-
-    // Format time
-    const formatTime = (dateString) => {
-        return new Date(dateString).toLocaleTimeString('hu-HU', { 
-            hour: '2-digit', 
-            minute: '2-digit' 
-        });
-    };
-
-    // Get status color
-    const getStatusColor = (status) => {
-        switch (status) {
-            case 'scheduled': return 'bg-blue-100 text-blue-700 border-blue-200';
-            case 'completed': return 'bg-green-100 text-green-700 border-green-200';
-            case 'canceled': return 'bg-red-100 text-red-700 border-red-200';
-            case 'no_show': return 'bg-gray-100 text-gray-700 border-gray-200';
-            default: return 'bg-gray-100 text-gray-700 border-gray-200';
-        }
-    };
-
-    // Get status text
-    const getStatusText = (status) => {
-        switch (status) {
-            case 'scheduled': return 'Foglalt';
-            case 'completed': return 'Teljesítve';
-            case 'canceled': return 'Törölve';
-            case 'no_show': return 'Nem jelent meg';
-            default: return status;
-        }
-    };
-
-    // Handle appointment click
-    const handleAppointmentClick = (appointment) => {
-        setSelectedAppointment(appointment);
-        setShowModal(true);
-    };
-
-    // Handle delete appointment
-    const handleDeleteAppointment = async () => {
-        if (!selectedAppointment) return;
-        
-        setDeleteLoading(true);
-        try {
-            const response = await authApi.delete(`/api/provider/calendar/appointments/${selectedAppointment.id}`);
-            const data = await response.json();
-            if (data.success) {
-                setShowModal(false);
-                setSelectedAppointment(null);
-                fetchAppointments();
-            }
-        } catch (error) {
-            console.error('Error deleting appointment:', error);
-        } finally {
-            setDeleteLoading(false);
-        }
-    };
-
-    // Days of week header
-    const daysOfWeek = ['H', 'K', 'Sze', 'Cs', 'P', 'Szo', 'V'];
-
-    // Format month name
-    const monthNames = ['Január', 'Február', 'Március', 'Április', 'Május', 'Június', 
-                        'Július', 'Augusztus', 'Szeptember', 'Október', 'November', 'December'];
-
-    const calendarDays = generateCalendarDays();
-    const todayAppointments = getAppointmentsForDate(selectedDate);
-    const todayTimeBlocks = getTimeBlocksForDate(selectedDate);
-
-    return (
-        <div className="space-y-4 sm:space-y-6">
-            <div className="flex justify-between items-center">
-                <h2 className="text-xl sm:text-2xl font-bold text-dark-blue">Naptár</h2>
-                <div className="flex gap-2">
-                    <button 
-                        onClick={() => { setSelectedTimeBlock(null); setShowTimeBlockModal(true); }}
-                        className="px-4 py-2 bg-gray-500 text-white rounded-xl font-medium hover:bg-gray-600 transition-colors shadow-md text-sm"
-                    >
-                        + Szünet
-                    </button>
-                    <button 
-                        onClick={handleOpenCreateModal}
-                        className="px-4 py-2 bg-dark-blue text-white rounded-xl font-medium hover:bg-blue-800 transition-colors shadow-md text-sm"
-                    >
-                        + Új Időpont
-                    </button>
-                </div>
-            </div>
-            
-            <div className="flex flex-col lg:flex-row gap-4 sm:gap-6">
-                {/* Calendar - Mini on mobile, larger on desktop */}
-                <div className="w-full lg:w-auto lg:min-w-[320px]">
-                    <div className="bg-white/40 backdrop-blur-md p-3 sm:p-6 rounded-2xl shadow-lg border border-white/50">
-                        {/* Month Navigation */}
-                        <div className="flex items-center justify-between mb-4">
-                            <button 
-                                onClick={prevMonth}
-                                className="p-1.5 sm:p-2 hover:bg-white/50 rounded-lg transition-colors"
-                            >
-                                <LeftArrowIcon className="w-4 h-4 sm:w-5 sm:h-5 text-dark-blue" />
-                            </button>
-                            <h3 className="text-base sm:text-lg font-bold text-dark-blue">
-                                {monthNames[currentDate.getMonth()]} {currentDate.getFullYear()}
-                            </h3>
-                            <button 
-                                onClick={nextMonth}
-                                className="p-1.5 sm:p-2 hover:bg-white/50 rounded-lg transition-colors"
-                            >
-                                <RightArrowIcon className="w-4 h-4 sm:w-5 sm:h-5 text-dark-blue" />
-                            </button>
-                        </div>
-
-                        {/* Days of week header */}
-                        <div className="grid grid-cols-7 gap-1 mb-2">
-                            {daysOfWeek.map((day, index) => (
-                                <div key={index} className="text-center text-[10px] sm:text-xs font-semibold text-gray-500 py-1 sm:py-2">
-                                    {day}
-                                </div>
-                            ))}
-                        </div>
-
-                        {/* Calendar grid */}
-                        <div className="grid grid-cols-7 gap-1">
-                            {calendarDays.map((date, index) => {
-                                const isToday = date && date.toDateString() === new Date().toDateString();
-                                const isSelected = date && date.toDateString() === selectedDate.toDateString();
-                                const bookingStatus = getBookingStatus(date);
-                                
-                                return (
-                                    <button
-                                        key={index}
-                                        onClick={() => date && setSelectedDate(date)}
-                                        disabled={!date}
-                                        className={`
-                                            aspect-square p-0.5 sm:p-1 rounded-lg text-xs sm:text-sm font-medium transition-all relative
-                                            ${!date ? 'cursor-default' : 'hover:bg-white/50 cursor-pointer'}
-                                            ${isToday ? 'ring-2 ring-dark-blue' : ''}
-                                            ${isSelected ? 'bg-dark-blue text-white' : 'text-gray-700'}
-                                        `}
-                                    >
-                                        {date?.getDate()}
-                                        {bookingStatus !== 'empty' && !isSelected && (
-                                            <div className={`absolute bottom-0.5 sm:bottom-1 left-1/2 -translate-x-1/2 w-1 h-1 sm:w-1.5 sm:h-1.5 rounded-full ${
-                                                bookingStatus === 'full' ? 'bg-red-500' : 'bg-green-500'
-                                            }`}></div>
-                                        )}
-                                    </button>
-                                );
-                            })}
-                        </div>
-                    </div>
-                </div>
-
-                {/* Day View - Schedule */}
-                <div className="flex-1">
-                    <div className="bg-white/40 backdrop-blur-md rounded-2xl shadow-lg border border-white/50 overflow-hidden">
-                        {/* Day header */}
-                        <div className="p-3 sm:p-4 border-b border-white/50 bg-white/20">
-                            <h3 className="text-base sm:text-lg font-bold text-dark-blue">
-                                {selectedDate.toLocaleDateString('hu-HU', { 
-                                    weekday: 'long', 
-                                    year: 'numeric', 
-                                    month: 'long', 
-                                    day: 'numeric' 
-                                })}
-                            </h3>
-                            <p className="text-xs sm:text-sm text-gray-600 mt-1">
-                                {todayAppointments.filter(a => a.status !== 'canceled').length} foglalás
-                                {todayTimeBlocks.length > 0 && ` · ${todayTimeBlocks.length} szünet`}
-                            </p>
-                        </div>
-
-                        {/* Timeline view */}
-                        <div className="max-h-100 sm:max-h-125 lg:max-h-150 overflow-y-auto">
-                            {loading ? (
-                                <div className="flex items-center justify-center py-12">
-                                    <div className="animate-spin rounded-full h-8 w-8 border-b-2 border-dark-blue"></div>
-                                </div>
-                            ) : (
-                                <div className="flex" style={{ height: `${TOTAL_MINUTES * MINUTES_PER_PIXEL}px` }}>
-                                    {/* Hour labels column */}
-                                    <div className="w-12 sm:w-16 shrink-0 border-r border-gray-100 relative">
-                                        {hours.map((hour, index) => (
-                                            <div 
-                                                key={hour} 
-                                                className="absolute left-0 right-0 pr-1 sm:pr-2 text-right -translate-y-1/2"
-                                                style={{ top: `${index * 60 * MINUTES_PER_PIXEL}px` }}
-                                            >
-                                                <span className="text-[10px] sm:text-xs font-medium text-gray-400">
-                                                    {hour.toString().padStart(2, '0')}:00
-                                                </span>
-                                            </div>
-                                        ))}
-                                    </div>
-                                    
-                                    {/* Timeline with appointments */}
-                                    <div 
-                                        className="flex-1 relative"
-                                    >
-                                        {/* Hour grid lines */}
-                                        {hours.map((hour) => (
-                                            <div 
-                                                key={hour}
-                                                className="absolute left-0 right-0 border-b-2 border-gray-300"
-                                                style={{ top: `${(hour - START_HOUR) * 60 * MINUTES_PER_PIXEL}px` }}
-                                            />
-                                        ))}
-                                        
-                                        {/* Half-hour grid lines */}
-                                        {hours.map((hour) => (
-                                            <div 
-                                                key={`half-${hour}`}
-                                                className="absolute left-0 right-0 border-b border-gray-200 border-dashed"
-                                                style={{ top: `${((hour - START_HOUR) * 60 + 30) * MINUTES_PER_PIXEL}px` }}
-                                            />
-                                        ))}
-                                        
-                                        {/* Appointments */}
-                                        {todayAppointments.length === 0 && todayTimeBlocks.length === 0 ? (
-                                            <div className="absolute inset-0 flex items-center justify-center">
-                                                <p className="text-gray-400 text-sm">Nincs foglalás erre a napra</p>
-                                            </div>
-                                        ) : (
-                                            <>
-                                            {/* Time Blocks - grayed background */}
-                                            {todayTimeBlocks.map((block, index) => {
-                                                const style = getTimeBlockStyle(block);
-                                                return (
-                                                    <button
-                                                        key={`block-${block.id}-${index}`}
-                                                        onClick={() => handleTimeBlockClick(block)}
-                                                        className="absolute left-0 right-0 rounded-lg transition-all overflow-hidden hover:opacity-80 active:scale-[0.99] cursor-pointer z-[1]"
-                                                        style={style}
-                                                    >
-                                                        <div className="w-full h-full bg-gray-300/60 border border-gray-400/40 border-dashed rounded-lg flex items-center justify-center px-2">
-                                                            <p className="text-[10px] sm:text-xs text-gray-600 font-medium truncate">
-                                                                {block.notes || 'Szünet'}
-                                                            </p>
-                                                        </div>
-                                                    </button>
-                                                );
-                                            })}
-                                            {/* Appointment blocks */}
-                                            {todayAppointments.map((apt) => {
-                                                const style = getAppointmentStyle(apt);
-                                                return (
-                                                    <button
-                                                        key={apt.id}
-                                                        onClick={() => handleAppointmentClick(apt)}
-                                                        className={`
-                                                            absolute left-1 right-1 sm:left-2 sm:right-2 
-                                                            px-1.5 py-0.5 sm:px-2 sm:py-1 rounded-lg border transition-all overflow-hidden
-                                                            hover:shadow-md hover:z-10 active:scale-[0.98]
-                                                            ${getStatusColor(apt.status)}
-                                                        `}
-                                                        style={style}
-                                                    >
-                                                        <div className="flex flex-col h-full justify-between">
-                                                            <div className="flex items-center justify-between gap-1">
-                                                                <p className="font-semibold text-[10px] sm:text-xs truncate">
-                                                                    {apt.user_name}
-                                                                </p>
-                                                                <span className="text-[10px] sm:text-xs font-medium whitespace-nowrap shrink-0">
-                                                                    {apt.price.toLocaleString()} Ft
-                                                                </span>
-                                                            </div>
-                                                            {apt.service_name && (
-                                                                <p className="text-[9px] sm:text-[10px] text-gray-600 truncate">
-                                                                    {apt.service_name}
-                                                                </p>
-                                                            )}
-                                                        </div>
-                                                    </button>
-                                                );
-                                            })}
-                                            </>
-                                        )}
-                                    </div>
-                                </div>
-                            )}
-                        </div>
-                    </div>
-                </div>
-            </div>
-
-            {/* Time Block Modal */}
-            <TimeBlockModal
-                isOpen={showTimeBlockModal}
-                onClose={() => { setShowTimeBlockModal(false); setSelectedTimeBlock(null); }}
-                onSaved={handleTimeBlockSaved}
-                block={selectedTimeBlock}
-                workingHours={workingHours}
-                selectedDate={selectedDate}
-            />
-
-            {/* Appointment Detail Modal */}
-            {showModal && selectedAppointment && createPortal(
-                <div className="fixed inset-0 z-100 flex items-center justify-center p-4">
-                    {/* Backdrop */}
-                    <div 
-                        className="absolute inset-0 bg-black/50 backdrop-blur-sm"
-                        onClick={() => setShowModal(false)}
-                    ></div>
-                    
-                    {/* Modal */}
-                    <div className="relative bg-white rounded-2xl shadow-2xl w-full max-w-md overflow-hidden animate-fade-in">
-                        {/* Header */}
-                        <div className={`p-4 sm:p-6 ${getStatusColor(selectedAppointment.status)}`}>
-                            <div className="flex items-start justify-between">
-                                <div>
-                                    <h3 className="text-lg sm:text-xl font-bold">
-                                        {selectedAppointment.user_name}
-                                    </h3>
-                                    {selectedAppointment.service_name && (
-                                        <p className="text-sm sm:text-base font-medium mt-0.5 opacity-90">
-                                            {selectedAppointment.service_name}
-                                        </p>
-                                    )}
-                                    <span className="inline-block mt-1 px-2 py-0.5 text-xs font-medium rounded-full bg-white/30">
-                                        {getStatusText(selectedAppointment.status)}
-                                    </span>
-                                </div>
-                                <button 
-                                    onClick={() => setShowModal(false)}
-                                    className="p-1 hover:bg-white/30 rounded-lg transition-colors"
-                                >
-                                    <CloseIcon className="w-5 h-5" />
-                                </button>
-                            </div>
-                        </div>
-
-                        {/* Content */}
-                        <div className="p-4 sm:p-6 space-y-4">
-                            {/* Time */}
-                            <div className="flex items-start gap-3">
-                                <div className="w-8 h-8 rounded-lg bg-blue-100 flex items-center justify-center shrink-0">
-                                    <HourIcon className="w-4 h-4 text-blue-600" />
-                                </div>
-                                <div>
-                                    <p className="text-xs text-gray-500">Időpont</p>
-                                    <p className="font-medium text-gray-900">
-                                        {new Date(selectedAppointment.appointment_start).toLocaleDateString('hu-HU', {
-                                            year: 'numeric',
-                                            month: 'long',
-                                            day: 'numeric'
-                                        })}
-                                    </p>
-                                    <p className="text-sm text-gray-600">
-                                        {formatTime(selectedAppointment.appointment_start)} - {formatTime(selectedAppointment.appointment_end)}
-                                    </p>
-                                </div>
-                            </div>
-
-                            {/* Contact */}
-                            <div className="flex items-start gap-3">
-                                <div className="w-8 h-8 rounded-lg bg-green-100 flex items-center justify-center shrink-0">
-                                    <UserIcon className="w-4 h-4 text-green-600" />
-                                </div>
-                                <div>
-                                    <p className="text-xs text-gray-500">Kapcsolat</p>
-                                    <p className="font-medium text-gray-900">{selectedAppointment.user_email}</p>
-                                    {selectedAppointment.user_phone && (
-                                        <p className="text-sm text-gray-600">{selectedAppointment.user_phone}</p>
-                                    )}
-                                </div>
-                            </div>
-
-                            {/* Price */}
-                            <div className="flex items-start gap-3">
-                                <div className="w-8 h-8 rounded-lg bg-yellow-100 flex items-center justify-center shrink-0">
-                                    <CurrencyIcon className="w-4 h-4 text-yellow-600" />
-                                </div>
-                                <div>
-                                    <p className="text-xs text-gray-500">Ár</p>
-                                    <p className="font-bold text-lg text-gray-900">
-                                        {selectedAppointment.price.toLocaleString()} Ft
-                                    </p>
-                                </div>
-                            </div>
-
-                            {/* Comment */}
-                            {selectedAppointment.comment && (
-                                <div className="flex items-start gap-3">
-                                    <div className="w-8 h-8 rounded-lg bg-purple-100 flex items-center justify-center shrink-0">
-                                        <ChatBubbleIcon className="w-4 h-4 text-purple-600" />
-                                    </div>
-                                    <div>
-                                        <p className="text-xs text-gray-500">Megjegyzés</p>
-                                        <p className="text-sm text-gray-700">{selectedAppointment.comment}</p>
-                                    </div>
-                                </div>
-                            )}
-                        </div>
-
-                        {/* Actions */}
-                        {selectedAppointment.status === 'scheduled' && (
-                            <div className="p-4 sm:p-6 border-t border-gray-100 bg-gray-50">
-                                <button
-                                    onClick={handleDeleteAppointment}
-                                    disabled={deleteLoading}
-                                    className="w-full py-2.5 px-4 bg-red-500 hover:bg-red-600 text-white font-medium rounded-xl transition-colors disabled:opacity-50 disabled:cursor-not-allowed flex items-center justify-center gap-2"
-                                >
-                                    {deleteLoading ? (
-                                        <div className="animate-spin rounded-full h-4 w-4 border-2 border-white border-t-transparent"></div>
-                                    ) : (
-                                        <>
-                                            <TrashIcon className="w-4 h-4" />
-                                            Foglalás törlése
-                                        </>
-                                    )}
-                                </button>
-                            </div>
-                        )}
-                    </div>
-                </div>,
-                document.body
-            )}
-
-            {/* Create Appointment Modal */}
-            {showCreateModal && createPortal(
-                <div className="fixed inset-0 z-[9999] flex items-center justify-center p-4">
-                    <div 
-                        className="fixed inset-0 bg-black/50 backdrop-blur-sm"
-                        onClick={() => setShowCreateModal(false)}
-                    ></div>
-                    <div className="relative bg-white rounded-2xl shadow-2xl w-full max-w-md my-8 max-h-[90vh] overflow-y-auto">
-                        <div className="p-6 border-b border-gray-100">
-                            <div className="flex items-center justify-between">
-                                <h3 className="text-xl font-bold text-dark-blue">Új Időpont Létrehozása</h3>
-                                <button 
-                                    onClick={() => setShowCreateModal(false)}
-                                    className="p-1 hover:bg-gray-100 rounded-lg transition-colors"
-                                >
-                                    <CloseIcon className="w-5 h-5" />
-                                </button>
-                            </div>
-                        </div>
-
-                        <div className="p-6 space-y-4">
-                            {/* Booking Type Toggle */}
-                            <div className="bg-gray-50 p-4 rounded-xl">
-                                <label className="block text-sm font-medium text-gray-700 mb-3">Foglalás típusa</label>
-                                <div className="flex gap-3">
-                                    <button
-                                        type="button"
-                                        onClick={() => setCreateFormData({ ...createFormData, is_guest: false, user_email: '', user_phone: '' })}
-                                        className={`flex-1 py-2 px-4 rounded-lg font-medium transition-all ${
-                                            !createFormData.is_guest 
-                                                ? 'bg-dark-blue text-white shadow-md' 
-                                                : 'bg-white text-gray-700 border border-gray-300 hover:border-dark-blue'
-                                        }`}
-                                    >
-                                        Regisztrált felhasználó
-                                    </button>
-                                    <button
-                                        type="button"
-                                        onClick={() => setCreateFormData({ ...createFormData, is_guest: true })}
-                                        className={`flex-1 py-2 px-4 rounded-lg font-medium transition-all ${
-                                            createFormData.is_guest 
-                                                ? 'bg-dark-blue text-white shadow-md' 
-                                                : 'bg-white text-gray-700 border border-gray-300 hover:border-dark-blue'
-                                        }`}
-                                    >
-                                        Vendég
-                                    </button>
-                                </div>
-                            </div>
-
-                            <div>
-                                <label className="block text-sm font-medium text-gray-700 mb-1">
-                                    Név *
-                                </label>
-                                <input
-                                    type="text"
-                                    value={createFormData.user_name}
-                                    onChange={(e) => setCreateFormData({ ...createFormData, user_name: e.target.value })}
-                                    className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-dark-blue focus:border-transparent"
-                                    placeholder="Kovács János"
-                                />
-                            </div>
-
-                            <div>
-                                <label className="block text-sm font-medium text-gray-700 mb-1">
-                                    Email {!createFormData.is_guest ? '*' : ''}
-                                </label>
-                                <input
-                                    type="email"
-                                    value={createFormData.user_email}
-                                    onChange={(e) => setCreateFormData({ ...createFormData, user_email: e.target.value })}
-                                    className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-dark-blue focus:border-transparent"
-                                    placeholder="pelda@email.hu"
-                                />
-                                {!createFormData.is_guest && (
-                                    <p className="text-xs text-gray-500 mt-1">Meglévő felhasználó email címe vagy új létrehozásához</p>
-                                )}
-                            </div>
-
-                            <div>
-                                <label className="block text-sm font-medium text-gray-700 mb-1">
-                                    Telefonszám {createFormData.is_guest && !createFormData.user_email ? '*' : ''}
-                                </label>
-                                <input
-                                    type="tel"
-                                    value={createFormData.user_phone}
-                                    onChange={(e) => setCreateFormData({ ...createFormData, user_phone: e.target.value })}
-                                    className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-dark-blue focus:border-transparent"
-                                    placeholder="+36 20 123 4567"
-                                />
-                            </div>
-
-                            <div>
-                                <label className="block text-sm font-medium text-gray-700 mb-1">Szolgáltatás *</label>
-                                <select
-                                    value={createFormData.service_id}
-                                    onChange={(e) => setCreateFormData({ ...createFormData, service_id: e.target.value })}
-                                    className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-dark-blue focus:border-transparent"
-                                >
-                                    <option value="">Válassz szolgáltatást...</option>
-                                    {services.map(service => (
-                                        <option key={service.id} value={service.id}>
-                                            {service.name} - {service.duration_minutes} perc - {service.price.toLocaleString()} Ft
-                                        </option>
-                                    ))}
-                                </select>
-                            </div>
-
-                            <div className="grid grid-cols-2 gap-4">
-                                <div>
-                                    <label className="block text-sm font-medium text-gray-700 mb-1">Dátum *</label>
-                                    <input
-                                        type="date"
-                                        value={createFormData.appointment_date}
-                                        onChange={(e) => setCreateFormData({ ...createFormData, appointment_date: e.target.value })}
-                                        className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-dark-blue focus:border-transparent"
-                                    />
-                                </div>
-                                <div>
-                                    <label className="block text-sm font-medium text-gray-700 mb-1">Időpont *</label>
-                                    <TimeSpinner
-                                        value={createFormData.appointment_time}
-                                        onChange={(timeString) => setCreateFormData({ ...createFormData, appointment_time: timeString })}
-                                        minHour={workingHours.openingHour}
-                                        maxHour={workingHours.closingHour}
-                                    />
-                                </div>
-                            </div>
-
-                            <div>
-                                <label className="block text-sm font-medium text-gray-700 mb-1">Megjegyzés</label>
-                                <textarea
-                                    value={createFormData.comment}
-                                    onChange={(e) => setCreateFormData({ ...createFormData, comment: e.target.value })}
-                                    className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-dark-blue focus:border-transparent resize-none"
-                                    rows={3}
-                                    placeholder="Speciális kérések, megjegyzések..."
-                                />
-                            </div>
-                        </div>
-
-                        <div className="p-6 border-t border-gray-100 bg-gray-50 flex gap-3">
-                            <button
-                                onClick={() => setShowCreateModal(false)}
-                                className="flex-1 py-2.5 px-4 border border-gray-300 text-gray-700 font-medium rounded-xl hover:bg-gray-100 transition-colors"
-                            >
-                                Mégse
-                            </button>
-                            <button
-                                onClick={handleCreateAppointment}
-                                disabled={saving}
-                                className="flex-1 py-2.5 px-4 bg-dark-blue text-white font-medium rounded-xl hover:bg-blue-800 transition-colors disabled:opacity-50 disabled:cursor-not-allowed flex items-center justify-center gap-2"
-                            >
-                                {saving ? (
-                                    <div className="animate-spin rounded-full h-4 w-4 border-2 border-white border-t-transparent"></div>
-                                ) : (
-                                    'Létrehozás'
-                                )}
-                            </button>
-                        </div>
-                    </div>
-                </div>,
-                document.body
-            )}
-        </div>
-    );
-};
-
-const ServicesSection = () => {
-    const [services, setServices] = useState([]);
-    const [loading, setLoading] = useState(true);
-    const [error, setError] = useState(null);
-    const [showModal, setShowModal] = useState(false);
-    const [editingService, setEditingService] = useState(null);
-    const [deleteConfirm, setDeleteConfirm] = useState(null);
-    const [saving, setSaving] = useState(false);
-    const [deleting, setDeleting] = useState(false);
-    const [formData, setFormData] = useState({
-        name: '',
-        description: '',
-        duration_minutes: 30,
-        price: 0,
-        status: 'available'
-    });
-
-    useEffect(() => {
-        fetchServices();
-    }, []);
-
-    const fetchServices = async () => {
-        try {
-            setLoading(true);
-            const response = await authApi.get('/api/provider/calendar/services');
-            const data = await response.json();
-            if (data.success) {
-                setServices(data.services);
-            } else {
-                setError('Nem sikerült betölteni a szolgáltatásokat');
-            }
-        } catch (error) {
-            console.error('Fetch services error:', error);
-            setError('Hiba történt a szolgáltatások betöltésekor');
-        } finally {
-            setLoading(false);
-        }
-    };
-
-    const formatPrice = (price) => {
-        return new Intl.NumberFormat('hu-HU').format(price);
-    };
-
-    const handleOpenModal = (service = null) => {
-        if (service) {
-            setEditingService(service);
-            setFormData({
-                name: service.name,
-                description: service.description || '',
-                duration_minutes: service.duration_minutes,
-                price: service.price,
-                status: service.status
-            });
-        } else {
-            setEditingService(null);
-            setFormData({
-                name: '',
-                description: '',
-                duration_minutes: 30,
-                price: 0,
-                status: 'available'
-            });
-        }
-        setShowModal(true);
-    };
-
-    const handleCloseModal = () => {
-        setShowModal(false);
-        setEditingService(null);
-        setFormData({
-            name: '',
-            description: '',
-            duration_minutes: 30,
-            price: 0,
-            status: 'available'
-        });
-    };
-
-    const handleSave = async () => {
-        if (!formData.name.trim()) {
-            showToast('A szolgáltatás neve kötelező!', 'warning');
-            return;
-        }
-        if (formData.duration_minutes < 5) {
-            showToast('A minimum időtartam 5 perc!', 'warning');
-            return;
-        }
-        if (formData.price < 0) {
-            showToast('Az ár nem lehet negatív!', 'warning');
-            return;
-        }
-
-        try {
-            setSaving(true);
-            let response;
-            
-            if (editingService) {
-                response = await authApi.put(`/api/provider/calendar/services/${editingService.id}`, formData);
-            } else {
-                response = await authApi.post('/api/provider/calendar/services', formData);
-            }
-
-            const data = await response.json();
-            if (data.success) {
-                handleCloseModal();
-                fetchServices();
-            } else {
-                showToast(data.message || 'Hiba történt a mentés során', 'error');
-            }
-        } catch (error) {
-            console.error('Save service error:', error);
-            showToast('Hiba történt a mentés során', 'error');
-        } finally {
-            setSaving(false);
-        }
-    };
-
-    const handleDelete = async (serviceId) => {
-        try {
-            setDeleting(true);
-            const response = await authApi.delete(`/api/provider/calendar/services/${serviceId}`);
-            const data = await response.json();
-            
-            if (data.success) {
-                setDeleteConfirm(null);
-                fetchServices();
-            } else {
-                showToast(data.message || 'Hiba történt a törlés során', 'error');
-            }
-        } catch (error) {
-            console.error('Delete service error:', error);
-            showToast('Hiba történt a törlés során', 'error');
-        } finally {
-            setDeleting(false);
-        }
-    };
-
-    return (
-        <div className="space-y-6">
-            <div className="flex justify-between items-center">
-                <h2 className="text-2xl font-bold text-dark-blue">Szolgáltatások Kezelése</h2>
-                <button 
-                    onClick={() => handleOpenModal()}
-                    className="px-4 py-2 bg-dark-blue text-white rounded-xl font-medium hover:bg-blue-800 transition-colors shadow-md"
-                >
-                    + Új Szolgáltatás
-                </button>
-            </div>
-
-            {loading ? (
-                <div className="flex items-center justify-center py-12">
-                    <div className="animate-spin rounded-full h-12 w-12 border-b-2 border-dark-blue"></div>
-                </div>
-            ) : error ? (
-                <div className="bg-red-50 border border-red-200 rounded-xl p-4 text-red-700">
-                    {error}
-                </div>
-            ) : services.length === 0 ? (
-                <div className="bg-white/40 backdrop-blur-md p-12 rounded-2xl shadow-lg border border-white/50 text-center">
-                    <ServicesIcon className="w-16 h-16 mx-auto text-gray-400 mb-4" />
-                    <p className="text-gray-600 text-lg">Még nincsenek szolgáltatások</p>
-                    <p className="text-gray-500 text-sm mt-2">Kattints a "+ Új Szolgáltatás" gombra a kezdéshez</p>
-                </div>
-            ) : (
-                <div className="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-3 gap-6">
-                    {services.map((service) => (
-                        <div key={service.id} className="group bg-white/40 backdrop-blur-md p-6 rounded-2xl shadow-lg border border-white/50 hover:border-white/80 transition-all hover:-translate-y-1">
-                            <div className="flex justify-between items-start mb-4">
-                                <div className="w-12 h-12 bg-light-blue rounded-lg flex items-center justify-center text-dark-blue">
-                                    <ServicesIcon />
-                                </div>
-                                <div className="flex gap-2 opacity-0 group-hover:opacity-100 transition-opacity">
-                                    <button 
-                                        onClick={() => handleOpenModal(service)}
-                                        className="p-1 hover:bg-white/50 rounded text-blue-600"
-                                    >
-                                        <PencilIcon className="w-5 h-5" />
-                                    </button>
-                                    <button 
-                                        onClick={() => setDeleteConfirm(service)}
-                                        className="p-1 hover:bg-white/50 rounded text-red-500"
-                                    >
-                                        <TrashIcon className="w-5 h-5" />
-                                    </button>
-                                </div>
-                            </div>
-                            <h3 className="font-bold text-lg text-gray-900">{service.name}</h3>
-                            {service.description && (
-                                <p className="text-gray-600 text-sm mt-1 line-clamp-2">{service.description}</p>
-                            )}
-                            <div className="flex items-center justify-between mt-3 pt-3 border-t border-gray-200">
-                                <span className="text-gray-500 text-sm">{service.duration_minutes} perc</span>
-                                <span className="font-bold text-dark-blue">{formatPrice(service.price)} Ft</span>
-                            </div>
-                            {service.status !== 'available' && (
-                                <div className="mt-2">
-                                    <span className="inline-block px-2 py-1 bg-yellow-100 text-yellow-700 rounded-full text-xs font-medium">
-                                        Nem elérhető
-                                    </span>
-                                </div>
-                            )}
-                        </div>
-                    ))}
-                </div>
-            )}
-
-            {/* Add/Edit Service Modal */}
-            {showModal && createPortal(
-                <div className="fixed inset-0 z-[9999] flex items-center justify-center p-4 overflow-y-auto">
-                    <div 
-                        className="fixed inset-0 bg-black/50 backdrop-blur-sm"
-                        onClick={handleCloseModal}
-                    ></div>
-                    <div className="relative bg-white rounded-2xl shadow-2xl w-full max-w-md my-8 max-h-[calc(100vh-4rem)] overflow-y-auto">
-                        <div className="p-6 border-b border-gray-100">
-                            <div className="flex items-center justify-between">
-                                <h3 className="text-xl font-bold text-dark-blue">
-                                    {editingService ? 'Szolgáltatás Szerkesztése' : 'Új Szolgáltatás'}
-                                </h3>
-                                <button 
-                                    onClick={handleCloseModal}
-                                    className="p-1 hover:bg-gray-100 rounded-lg transition-colors"
-                                >
-                                    <CloseIcon className="w-5 h-5" />
-                                </button>
-                            </div>
-                        </div>
-
-                        <div className="p-6 space-y-4">
-                            <div>
-                                <label className="block text-sm font-medium text-gray-700 mb-1">Név *</label>
-                                <input
-                                    type="text"
-                                    value={formData.name}
-                                    onChange={(e) => setFormData({ ...formData, name: e.target.value })}
-                                    className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-dark-blue focus:border-transparent"
-                                    placeholder="pl. Férfi Hajvágás"
-                                />
-                            </div>
-
-                            <div>
-                                <label className="block text-sm font-medium text-gray-700 mb-1">Leírás</label>
-                                <textarea
-                                    value={formData.description}
-                                    onChange={(e) => setFormData({ ...formData, description: e.target.value })}
-                                    className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-dark-blue focus:border-transparent resize-none"
-                                    rows={3}
-                                    placeholder="Rövid leírás a szolgáltatásról..."
-                                />
-                            </div>
-
-                            <div className="grid grid-cols-2 gap-4">
-                                <div>
-                                    <label className="block text-sm font-medium text-gray-700 mb-1">Időtartam (perc) *</label>
-                                    <input
-                                        type="number"
-                                        value={formData.duration_minutes}
-                                        onChange={(e) => setFormData({ ...formData, duration_minutes: parseInt(e.target.value) || 0 })}
-                                        className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-dark-blue focus:border-transparent"
-                                        min="5"
-                                        max="480"
-                                    />
-                                </div>
-                                <div>
-                                    <label className="block text-sm font-medium text-gray-700 mb-1">Ár (Ft) *</label>
-                                    <input
-                                        type="number"
-                                        value={formData.price}
-                                        onChange={(e) => setFormData({ ...formData, price: parseInt(e.target.value) || 0 })}
-                                        className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-dark-blue focus:border-transparent"
-                                        min="0"
-                                    />
-                                </div>
-                            </div>
-
-                            <div>
-                                <label className="block text-sm font-medium text-gray-700 mb-1">Státusz</label>
-                                <select
-                                    value={formData.status}
-                                    onChange={(e) => setFormData({ ...formData, status: e.target.value })}
-                                    className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-dark-blue focus:border-transparent"
-                                >
-                                    <option value="available">Elérhető</option>
-                                    <option value="unavailable">Nem elérhető</option>
-                                </select>
-                            </div>
-                        </div>
-
-                        <div className="p-6 border-t border-gray-100 bg-gray-50 flex gap-3">
-                            <button
-                                onClick={handleCloseModal}
-                                className="flex-1 py-2.5 px-4 border border-gray-300 text-gray-700 font-medium rounded-xl hover:bg-gray-100 transition-colors"
-                            >
-                                Mégse
-                            </button>
-                            <button
-                                onClick={handleSave}
-                                disabled={saving}
-                                className="flex-1 py-2.5 px-4 bg-dark-blue text-white font-medium rounded-xl hover:bg-blue-800 transition-colors disabled:opacity-50 disabled:cursor-not-allowed flex items-center justify-center gap-2"
-                            >
-                                {saving ? (
-                                    <div className="animate-spin rounded-full h-4 w-4 border-2 border-white border-t-transparent"></div>
-                                ) : (
-                                    editingService ? 'Mentés' : 'Létrehozás'
-                                )}
-                            </button>
-                        </div>
-                    </div>
-                </div>,
-                document.body
-            )}
-
-            {/* Delete Confirmation Modal */}
-            {deleteConfirm && createPortal(
-                <div className="fixed inset-0 z-[9999] flex items-center justify-center p-4 overflow-y-auto">
-                    <div 
-                        className="fixed inset-0 bg-black/50 backdrop-blur-sm"
-                        onClick={() => setDeleteConfirm(null)}
-                    ></div>
-                    <div className="relative bg-white rounded-2xl shadow-2xl w-full max-w-sm my-8">
-                        <div className="p-6">
-                            <div className="w-12 h-12 rounded-full bg-red-100 flex items-center justify-center mx-auto mb-4">
-                                <WarningIcon className="w-6 h-6 text-red-600" />
-                            </div>
-                            <h3 className="text-lg font-bold text-center text-gray-900 mb-2">Szolgáltatás Törlése</h3>
-                            <p className="text-gray-600 text-center text-sm">
-                                Biztosan törölni szeretnéd a <strong>"{deleteConfirm.name}"</strong> szolgáltatást? Ez a művelet nem vonható vissza.
-                            </p>
-                        </div>
-                        <div className="p-4 border-t border-gray-100 bg-gray-50 flex gap-3">
-                            <button
-                                onClick={() => setDeleteConfirm(null)}
-                                className="flex-1 py-2.5 px-4 border border-gray-300 text-gray-700 font-medium rounded-xl hover:bg-gray-100 transition-colors"
-                            >
-                                Mégse
-                            </button>
-                            <button
-                                onClick={() => handleDelete(deleteConfirm.id)}
-                                disabled={deleting}
-                                className="flex-1 py-2.5 px-4 bg-red-500 text-white font-medium rounded-xl hover:bg-red-600 transition-colors disabled:opacity-50 disabled:cursor-not-allowed flex items-center justify-center gap-2"
-                            >
-                                {deleting ? (
-                                    <div className="animate-spin rounded-full h-4 w-4 border-2 border-white border-t-transparent"></div>
-                                ) : (
-                                    'Törlés'
-                                )}
-                            </button>
-                        </div>
-                    </div>
-                </div>,
-                document.body
-            )}
-        </div>
-    );
-};
-
-const NavButton = ({ activeTab, tabId, label, icon, onClick, isMobile }) => (
-    <button
-        onClick={() => onClick(tabId)}
-        className={`flex items-center gap-3 p-3 rounded-xl transition-all duration-300 relative overflow-hidden
-            ${activeTab === tabId
-                ? 'bg-dark-blue text-white shadow-lg scale-105'
-                : 'hover:bg-white/40 text-gray-700 hover:text-dark-blue'
-            }
-            ${isMobile ? 'flex-col text-[10px] gap-1 p-2 w-full justify-center' : 'w-full'}
-        `}
-    >
-        <div className={`transition-transform duration-300 ${activeTab === tabId && !isMobile ? 'translate-x-1' : ''}`}>
-             {icon}
-        </div>
-        <span className={`${isMobile ? 'font-medium' : 'font-semibold'}`}>{label}</span>
-        
-        {/* Active Indicator Glow */}
-        {activeTab === tabId && (
-            <div className="absolute inset-0 bg-white/10 opacity-50 blur-md rounded-xl"></div>
-        )}
-    </button>
-);
-
-const UserDropdown = ({ isOpen, onLogout, onProfileEdit, providerProfile }) => {
-    if (!isOpen) return null;
-    
-    return (
-        <div className="absolute top-14 right-4 w-52 bg-white/60 backdrop-blur-md border border-white/50 rounded-2xl shadow-xl overflow-hidden animate-fade-in z-50">
-            <div className="p-4 border-b border-gray-100">
-                <p className="text-sm font-bold text-gray-900">{providerProfile?.name || 'Szolgáltató'}</p>
-                <p className="text-xs text-gray-500 truncate">{providerProfile?.email || ''}</p>
-            </div>
-            <div className="p-2 space-y-1">
-                <button onClick={onProfileEdit} className="w-full text-left px-3 py-2 text-sm text-gray-700 hover:bg-white/50 rounded-lg transition-colors flex items-center gap-2">
-                    <UserIcon className="w-4 h-4" />
-                    Profil Szerkesztése
-                </button>
-                <button className="w-full text-left px-3 py-2 text-sm text-gray-700 hover:bg-white/50 rounded-lg transition-colors flex items-center gap-2">
-                    <SettingsIcon className="w-4 h-4" />
-                    Beállítások
-                </button>
-                <div className="border-t border-gray-100 my-1"></div>
-                <button 
-                    onClick={onLogout}
-                    className="w-full text-left px-3 py-2 text-sm text-red-600 hover:bg-red-50 rounded-lg transition-colors flex items-center gap-2"
-                >
-                    <LogoutIcon className="w-4 h-4" />
-                    Kijelentkezés
-                </button>
-            </div>
-        </div>
-    );
-};
+import OverviewSection from './provdashcomponents/OverviewSection';
+import CalendarSection from './provdashcomponents/CalendarSection/CalendarSection';
+import ServicesSection from './provdashcomponents/ServicesSection/ServicesSection';
+import NavButton from './provdashcomponents/NavButton';
+import UserDropdown from './provdashcomponents/UserDropdown';
+import ProfileModal from './provdashcomponents/ProfileModal';
+import PasswordModal from './provdashcomponents/PasswordModal';
 
 export default function ProvDash() {
     const [activeTab, setActiveTab] = useState('overview');
@@ -1561,7 +102,7 @@ export default function ProvDash() {
             } else {
                 setProfileError(data.message || 'Hiba történt a mentés során');
             }
-        } catch (error) {
+        } catch {
             setProfileError('Hiba történt a mentés során');
         } finally {
             setProfileSaving(false);
@@ -1589,7 +130,7 @@ export default function ProvDash() {
             } else {
                 setPictureError(data.message || 'Hiba történt a kép feltöltésekor');
             }
-        } catch (error) {
+        } catch {
             setPictureError('Hiba történt a kép feltöltésekor');
         } finally {
             setPictureUploading(false);
@@ -1621,7 +162,7 @@ export default function ProvDash() {
             } else {
                 setPasswordError(data.message || 'Hiba történt a jelszó módosítása során');
             }
-        } catch (error) {
+        } catch {
             setPasswordError('Hiba történt a jelszó módosítása során');
         } finally {
             setPasswordSaving(false);
@@ -1644,6 +185,7 @@ export default function ProvDash() {
     const handleLogout = async () => {
         await logout();
         localStorage.removeItem('accessToken');
+        logout();
         setIsAuthenticated(false);
         navigate('/');
     };
@@ -1678,7 +220,7 @@ export default function ProvDash() {
                         {providerProfile?.profile_picture_url ? (
                             <img src={`${apiUrl}${providerProfile.profile_picture_url}`} alt="Profil" className="w-10 h-10 rounded-full object-cover shadow-lg hover:shadow-xl hover:scale-105 transition-all border-2 border-white/50 cursor-pointer" />
                         ) : (
-                            <div className="w-10 h-10 rounded-full bg-linear-to-br from-dark-blue to-blue-500 text-white flex items-center justify-center shadow-lg hover:shadow-xl hover:scale-105 transition-all text-sm font-bold border-2 border-white/50 cursor-pointer">
+                            <div className="w-10 h-10 rounded-full bg-gradient-to-br from-dark-blue to-blue-500 text-white flex items-center justify-center shadow-lg hover:shadow-xl hover:scale-105 transition-all text-sm font-bold border-2 border-white/50 cursor-pointer">
                                 {getProviderInitials(providerProfile?.name || user?.name)}
                             </div>
                         )}
@@ -1786,133 +328,32 @@ export default function ProvDash() {
                 </nav>
             </div>
 
-            {/* Profile Edit Modal */}
-            {showProfileModal && createPortal(
-                <div className="fixed inset-0 z-[9999] flex items-center justify-center p-4 overflow-y-auto">
-                    <div className="fixed inset-0 bg-black/50 backdrop-blur-sm" onClick={() => setShowProfileModal(false)}></div>
-                    <div className="relative bg-white rounded-2xl shadow-2xl w-full max-w-md my-8 max-h-[calc(100vh-4rem)] overflow-y-auto">
-                        <div className="p-6 border-b border-gray-100">
-                            <div className="flex items-center justify-between">
-                                <h3 className="text-xl font-bold text-dark-blue">Profil szerkesztése</h3>
-                                <button onClick={() => setShowProfileModal(false)} className="p-1 hover:bg-gray-100 rounded-lg transition-colors">
-                                    <CloseIcon className="w-5 h-5" />
-                                </button>
-                            </div>
-                        </div>
-                        <div className="p-6 space-y-5">
-                            {profileSuccess && (
-                                <div className="p-3 bg-green-50 border border-green-200 rounded-lg text-green-700 text-sm flex items-center gap-2">
-                                    <CheckCircleIcon className="w-5 h-5 shrink-0" />
-                                    {profileSuccess}
-                                </div>
-                            )}
-                            {/* Avatar section */}
-                            <div className="flex flex-col items-center gap-3">
-                                <div className="relative">
-                                    {providerProfile?.profile_picture_url ? (
-                                        <img src={`${apiUrl}${providerProfile.profile_picture_url}`} alt="Profil" className="w-24 h-24 rounded-full object-cover border-4 border-white shadow-lg" />
-                                    ) : (
-                                        <div className="w-24 h-24 rounded-full bg-gradient-to-br from-dark-blue to-blue-500 text-white flex items-center justify-center text-2xl font-bold border-4 border-white shadow-lg">
-                                            {getProviderInitials(providerProfile?.name)}
-                                        </div>
-                                    )}
-                                    {pictureUploading && (
-                                        <div className="absolute inset-0 flex items-center justify-center bg-black/40 rounded-full">
-                                            <div className="animate-spin rounded-full h-8 w-8 border-2 border-white border-t-transparent"></div>
-                                        </div>
-                                    )}
-                                </div>
-                                <label className={`cursor-pointer px-4 py-2 bg-gray-100 hover:bg-gray-200 text-gray-700 text-sm font-medium rounded-lg transition-colors ${pictureUploading ? 'opacity-50 pointer-events-none' : ''}`}>
-                                    {pictureUploading ? 'Feltöltés...' : 'Kép módosítása'}
-                                    <input type="file" accept="image/jpeg,image/png,image/webp" className="hidden" onChange={handleProviderPictureUpload} disabled={pictureUploading} />
-                                </label>
-                                {pictureError && <p className="text-red-500 text-xs">{pictureError}</p>}
-                                <p className="text-xs text-gray-400">Max 5MB • JPG, PNG, WebP</p>
-                            </div>
-                            {profileError && (<div className="p-3 bg-red-50 border border-red-200 rounded-lg text-red-700 text-sm">{profileError}</div>)}
-                            <div>
-                                <label className="block text-sm font-medium text-gray-700 mb-1">Név *</label>
-                                <input type="text" value={profileFormData.name} onChange={(e) => setProfileFormData({...profileFormData, name: e.target.value})} className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-dark-blue focus:border-transparent" placeholder="Teljes név" disabled={profileSaving} />
-                            </div>
-                            <div>
-                                <label className="block text-sm font-medium text-gray-700 mb-1">Email</label>
-                                <input type="email" value={providerProfile?.email || ''} className="w-full px-3 py-2 border border-gray-200 rounded-lg bg-gray-50 text-gray-500 cursor-not-allowed" disabled />
-                                <p className="text-xs text-gray-400 mt-1">Az email cím nem módosítható</p>
-                            </div>
-                            <div>
-                                <label className="block text-sm font-medium text-gray-700 mb-1">Telefonszám *</label>
-                                <input type="tel" value={profileFormData.phone} onChange={(e) => setProfileFormData({...profileFormData, phone: e.target.value})} className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-dark-blue focus:border-transparent" placeholder="+36 20 123 4567" disabled={profileSaving} />
-                            </div>
-                            <div>
-                                <label className="block text-sm font-medium text-gray-700 mb-1">Bemutatkozás</label>
-                                <textarea value={profileFormData.description} onChange={(e) => setProfileFormData({...profileFormData, description: e.target.value})} className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-dark-blue focus:border-transparent resize-none" rows={3} placeholder="Rövid bemutatkozás..." disabled={profileSaving} />
-                            </div>
-                            <div className="pt-2 border-t border-gray-200">
-                                <button onClick={openPasswordModal} className="w-full text-left px-3 py-3 text-sm font-medium text-gray-700 hover:bg-gray-50 rounded-lg transition-colors flex items-center justify-between">
-                                    <span className="flex items-center gap-2">
-                                        <LockIcon className="w-4 h-4" />
-                                        Jelszó módosítása
-                                    </span>
-                                    <RightArrowIcon className="w-4 h-4" />
-                                </button>
-                            </div>
-                        </div>
-                        <div className="p-6 border-t border-gray-100 bg-gray-50 flex gap-3">
-                            <button onClick={() => setShowProfileModal(false)} className="flex-1 py-2.5 px-4 border border-gray-300 text-gray-700 font-medium rounded-xl hover:bg-gray-100 transition-colors" disabled={profileSaving}>Mégse</button>
-                            <button onClick={handleProviderProfileSave} disabled={profileSaving} className="flex-1 py-2.5 px-4 bg-dark-blue text-white font-medium rounded-xl hover:bg-blue-800 transition-colors disabled:opacity-50 disabled:cursor-not-allowed flex items-center justify-center gap-2">
-                                {profileSaving ? <div className="animate-spin rounded-full h-4 w-4 border-2 border-white border-t-transparent"></div> : 'Mentés'}
-                            </button>
-                        </div>
-                    </div>
-                </div>,
-                document.body
-            )}
+            <ProfileModal 
+                isOpen={showProfileModal}
+                onClose={() => setShowProfileModal(false)}
+                providerProfile={providerProfile}
+                profileFormData={profileFormData}
+                setProfileFormData={setProfileFormData}
+                profileSaving={profileSaving}
+                profileError={profileError}
+                profileSuccess={profileSuccess}
+                pictureUploading={pictureUploading}
+                pictureError={pictureError}
+                onSave={handleProviderProfileSave}
+                onPictureUpload={handleProviderPictureUpload}
+                onPasswordModalOpen={openPasswordModal}
+            />
 
-            {/* Password Change Modal */}
-            {showPasswordModal && createPortal(
-                <div className="fixed inset-0 z-[10000] flex items-center justify-center p-4 overflow-y-auto">
-                    <div className="fixed inset-0 bg-black/50 backdrop-blur-sm" onClick={() => setShowPasswordModal(false)}></div>
-                    <div className="relative bg-white rounded-2xl shadow-2xl w-full max-w-md my-8 max-h-[calc(100vh-4rem)] overflow-y-auto">
-                        <div className="p-6 border-b border-gray-100">
-                            <div className="flex items-center justify-between">
-                                <h3 className="text-xl font-bold text-dark-blue">Jelszó módosítása</h3>
-                                <button onClick={() => setShowPasswordModal(false)} className="p-1 hover:bg-gray-100 rounded-lg transition-colors">
-                                    <CloseIcon className="w-5 h-5" />
-                                </button>
-                            </div>
-                        </div>
-                        <div className="p-6 space-y-4">
-                            {passwordError && (<div className="p-3 bg-red-50 border border-red-200 rounded-lg text-red-700 text-sm">{passwordError}</div>)}
-                            {passwordSuccess && (
-                                <div className="p-3 bg-green-50 border border-green-200 rounded-lg text-green-700 text-sm flex items-center gap-2">
-                                    <CheckCircleIcon className="w-5 h-5 shrink-0" />
-                                    {passwordSuccess}
-                                </div>
-                            )}
-                            <div>
-                                <label className="block text-sm font-medium text-gray-700 mb-1">Jelenlegi jelszó *</label>
-                                <input type="password" value={passwordFormData.currentPassword} onChange={(e) => setPasswordFormData({...passwordFormData, currentPassword: e.target.value})} className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-dark-blue focus:border-transparent" placeholder="••••••••" disabled={passwordSaving} />
-                            </div>
-                            <div>
-                                <label className="block text-sm font-medium text-gray-700 mb-1">Új jelszó *</label>
-                                <input type="password" value={passwordFormData.newPassword} onChange={(e) => setPasswordFormData({...passwordFormData, newPassword: e.target.value})} className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-dark-blue focus:border-transparent" placeholder="••••••••" disabled={passwordSaving} />
-                                <p className="text-xs text-gray-500 mt-1">Legalább 6 karakter hosszú legyen</p>
-                            </div>
-                            <div>
-                                <label className="block text-sm font-medium text-gray-700 mb-1">Új jelszó megerősítése *</label>
-                                <input type="password" value={passwordFormData.confirmPassword} onChange={(e) => setPasswordFormData({...passwordFormData, confirmPassword: e.target.value})} className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-dark-blue focus:border-transparent" placeholder="••••••••" disabled={passwordSaving} />
-                            </div>
-                        </div>
-                        <div className="p-6 border-t border-gray-100 bg-gray-50 flex gap-3">
-                            <button onClick={() => setShowPasswordModal(false)} className="flex-1 py-2.5 px-4 border border-gray-300 text-gray-700 font-medium rounded-xl hover:bg-gray-100 transition-colors" disabled={passwordSaving}>Mégse</button>
-                            <button onClick={handleProviderPasswordChange} disabled={passwordSaving || !!passwordSuccess} className="flex-1 py-2.5 px-4 bg-dark-blue text-white font-medium rounded-xl hover:bg-blue-800 transition-colors disabled:opacity-50 disabled:cursor-not-allowed flex items-center justify-center gap-2">
-                                {passwordSaving ? <div className="animate-spin rounded-full h-4 w-4 border-2 border-white border-t-transparent"></div> : 'Jelszó módosítása'}
-                            </button>
-                        </div>
-                    </div>
-                </div>,
-                document.body
-            )}
+            <PasswordModal 
+                isOpen={showPasswordModal}
+                onClose={() => setShowPasswordModal(false)}
+                passwordFormData={passwordFormData}
+                setPasswordFormData={setPasswordFormData}
+                passwordSaving={passwordSaving}
+                passwordError={passwordError}
+                passwordSuccess={passwordSuccess}
+                onSave={handleProviderPasswordChange}
+            />
         </div>
     );
 }
